+++ lib/top_secret/error.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 module TopSecret
-  class Error < StandardError; end
+  class Error < StandardError
+    class MalformedLabel < StandardError; end
+  end
 end