+++ lib/top_secret/text/global_mapping.rb
@@ -50,8 +50,6 @@ module TopSecret
       # @param individual_key [Symbol] The individual key from a filter result
       # @return [Symbol] The global key with consistent numbering
       def generate_global_key(individual_key)
-        # TODO: This assumes labels are formatted consistently.
-        # We need to account for the following for the case where a label could begin with an "_"
         label_type = individual_key.to_s.rpartition("_").first
 
         label_counters[label_type] ||= 0